Move functionality to separate file
